+++ app.json
@@ -2,10 +2,10 @@
   "expo": {
     "name": "TrackQueen",
     "slug": "tq22",
-    "version": "1.2.2",
+    "version": "1.2.4",
     "orientation": "portrait",
     "description": "Learn more about your music.",
-    "scheme": "tq22",
+    "scheme": "com.jcanelis.tq22",
     "icon": "./assets/icon.png",
     "userInterfaceStyle": "automatic",
     "githubUrl": "https://github.com/jcanelis/trackqueen",
@@ -18,6 +18,7 @@
     "ios": {
       "icon": "./assets/app.icon",
       "bundleIdentifier": "com.jcanelis.tq22",
+      "associatedDomains": ["applinks:trackqueen.app"],
       "supportsTablet": true,
       "infoPlist": {
         "UIBackgroundModes": ["audio", "fetch", "audio", "fetch"],
@@ -74,7 +75,7 @@
       "expo-font",
       "expo-audio"
     ],
-    "runtimeVersion": "1.2.2",
+    "runtimeVersion": "1.2.4",
     "android": {
       "package": "com.jcanelis.tq22"
     }
