+++ eas.json
@@ -18,7 +18,6 @@
     },
     "development": {
       "extends": "base",
-      "environment": "development",
       "developmentClient": true,
       "distribution": "internal",
       "cache": {
@@ -27,7 +26,6 @@
     },
     "development-simulator": {
       "extends": "base",
-      "environment": "development",
       "developmentClient": true,
       "distribution": "internal",
       "ios": {
@@ -36,7 +34,6 @@
     },
     "preview": {
       "extends": "base",
-      "environment": "preview",
       "channel": "preview",
       "distribution": "internal",
       "cache": {
@@ -45,7 +42,6 @@
     },
     "production": {
       "autoIncrement": true,
-      "environment": "production",
       "extends": "base",
       "channel": "main",
       "cache": {
