+++ firebase/firebase.json
@@ -1,7 +1,7 @@
 {
   "functions": {
-    "source": "functions",
-    "predeploy": ["npm --prefix \"$RESOURCE_DIR\" run lint"]
+    "source": "functions"
+    //"predeploy": ["npm --prefix \"$RESOURCE_DIR\" run lint"]
   },
   "emulators": {
     "auth": {
