+++ firebase/functions/package.json
@@ -2,7 +2,7 @@
   "name": "functions",
   "description": "Cloud Functions for Firebase",
   "scripts": {
-    "lint": "eslint .",
+    "lint": "npx eslint .",
     "emulators": "firebase emulators:start",
     "serve": "firebase emulators:start --only functions",
     "shell": "firebase functions:shell",
@@ -11,26 +11,26 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "20"
+    "node": "24"
   },
   "main": "index.js",
   "dependencies": {
-    "@google-cloud/secret-manager": "^6.1.0",
-    "cors": "^2.8.5",
+    "@google-cloud/secret-manager": "^6.1.1",
+    "cors": "^2.8.6",
     "crypto": "^1.0.1",
-    "firebase-admin": "^13.5.0",
-    "firebase-functions": "^6.4.0",
-    "jsonwebtoken": "^9.0.2",
+    "firebase-admin": "^13.8.0",
+    "firebase-functions": "^7.2.5",
+    "jsonwebtoken": "^9.0.3",
     "querystring": "^0.2.1",
     "request": "^2.88.2"
   },
   "devDependencies": {
-    "eslint": "^9.36.0",
+    "eslint": "^10.2.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-react": "^7.37.5",
     "firebase-functions-test": "^3.4.1",
-    "firebase-tools": "^14.17.0",
-    "prettier": "^3.6.2"
+    "firebase-tools": "^15.14.0",
+    "prettier": "^3.8.3"
   },
   "private": true
 }
