+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "trackqueen",
-  "version": "1.2.2",
+  "version": "1.2.4",
   "description": "Learn more about your music.",
   "scripts": {
     "build": "npx eas-cli build --platform ios",
@@ -38,21 +38,21 @@
     "@shopify/flash-list": "2.0.2",
     "@tanstack/react-query": "^4.35.3",
     "crypto-js": "^4.2.0",
-    "expo": "~55.0.14",
-    "expo-asset": "~55.0.13",
+    "expo": "~55.0.15",
+    "expo-asset": "~55.0.15",
     "expo-audio": "~55.0.12",
-    "expo-auth-session": "~55.0.12",
+    "expo-auth-session": "~55.0.14",
     "expo-blur": "~55.0.13",
     "expo-build-properties": "~55.0.12",
-    "expo-crypto": "~55.0.13",
+    "expo-crypto": "~55.0.14",
     "expo-dev-client": "~55.0.23",
     "expo-file-system": "~55.0.16",
     "expo-font": "~55.0.6",
     "expo-glass-effect": "~55.0.10",
     "expo-image": "~55.0.8",
-    "expo-linking": "~55.0.11",
+    "expo-linking": "~55.0.13",
     "expo-secure-store": "~55.0.12",
-    "expo-splash-screen": "~55.0.16",
+    "expo-splash-screen": "~55.0.18",
     "expo-status-bar": "~55.0.5",
     "expo-symbols": "~55.0.7",
     "expo-system-ui": "~55.0.14",
@@ -82,7 +82,7 @@
     "eslint-plugin-react-hooks": "^7.0.1",
     "globals": "^17.3.0",
     "jest": "~29.7.0",
-    "jest-expo": "~55.0.14",
+    "jest-expo": "~55.0.16",
     "prettier": "^3.8.1",
     "react-test-renderer": "^19.2.4"
   },
